[JENKINS-72161] avoid inodes check takes agents online
also JENKINS-72160

Use an OfflineCause specific to the Inodes usage and check before taking
an agent online if the current cause is still the same. If not it wasn't
the plugin that took the agent offline, thus it needs to stay offline.

Use the cause as data object so we can hightlight the table cell
when agent is offline. This will also show a tooltip with the threshold.